Refactor some settings and jinja env.
